Factor out CI  into composable actions and reusable workflows(#11)
The previous CI system was a monolithic action with matrix dispatch. The new system factors out the build step into an action. Because GitHub does not separate badges for matrix builds :roll_eyes:  the matrix was split on nixpkgs version to track what versions are building in the readme (along with scheduled runs).

- Adds action for building the test repo with inputs for generation type, style, branch of repo, and branch of nixpkgs
- Add a reusable workflow for building all the packages with a set nix packages release
- Add workflows for 22.05, 22.11, 23.05, 23.11, unstable, and master
- Add badges for all workflows
- Add scheduled builds daily for active branches (23.05, 23.11, unstable, master)
- Add an on PR build that builds for all supported branches