Remove the on push to prevent extra CI runs and branch protection (#12)
Remove the on push to prevent extra CI runs and add a dummy job to allow for branch protection. CI was running when a PR got merged due to the on-push trigger in github ci. GH actions does not allow for workflows-based PR rejections on the repo level. Instead, this adds a dummy job that needs all the other jobs

- Remove the on push from the on_pr workflow 
- Add a it-all-good job to track the status of the build