Support custom `pythonPackages` in generation
The current generation only supports one version of `pythonPackages`
which is defaulted to the value of `final.python3Packages`. This
presents a problem for users that want to use nix-proto with custom
versions of `pythonPackages`.

The solution implemented here is to use
[`pythonPackagesExtensions`](https://github.com/NixOS/nixpkgs/pull/91850)
which is applied to all `pythonPackages` scopes. For now I will continue
to support direct access to the python generation but eventually it will
be removed in favor of all access through a `pythonPackages` scope.

This commit preserves the old python generation under a legacy name
e.g, grpcLegacy. There are also some quality of life changes:

- Add a default attribute to `generateOverlays'` that composes the overlays
- Add a throw with an explicit error if name given to
  `generateOverlays'` and `mkProtoDerivation` match.
- Deprecation warnings for using proto package type
- Clean up the `inherits` attribute in generation.nix
- Update flake to 24.05
- Fix flake lints