Rework CI to allow forks, local builds, and formatting (#25)
Through this #25 I reworked the CI to improve the legibility and
coverage. The main test data is in repo and the flake has been reworked
to user `flake-parts`. The old CI method is preserved as an API check
and example. This also supports forks as a place to build from.
Addtionally I added formatting for nix, yaml, and proto. Formatting is
not turned on yet to allow a tree wide format to be another PR.

* Small fix for C++ build for newer versions of protobuf