Fix package handling for overriden interpreters in cross builds
This replaces refering to the native dependencies by looking it up by
python version in `buildPackages` instead it uses `pythonOnBuildForHost`
to get the native package when cross compiling. This does increase the
memory usage for eval but not as much as using splicing which has a
performace problem for pythonPackages (see NixOS/nixpkgs#338231). This
is a potential alternative to #24.